@@ -6,12 +6,13 @@
     "homepage": "https://vrok.de",
     "license": "MIT",
     "require": {
-        "php": "^8.1",
-        "symfony/amqp-messenger": "^6.0.0|^7.0.0"
+        "php": "^8.3",
+        "symfony/amqp-messenger": "^7.0.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^3.40.2",
-        "phpunit/phpunit": "^10.5.1"
+        "friendsofphp/php-cs-fixer": "^3.89.1",
+        "phpunit/phpunit": "^12.4.2",
+        "rector/rector": "^2.2.7"
     },
     "autoload": {
         "psr-4": {
