@@ -1,5 +1,7 @@
 <?php
 
+/** @noinspection PhpUnhandledExceptionInspection */
+
 declare(strict_types=1);
 
 namespace Vrok\MessengerReply\Tests;
@@ -14,26 +16,23 @@ use Symfony\Component\Messenger\Middleware\StackMiddleware;
 use Symfony\Component\Messenger\Stamp\HandledStamp;
 use Symfony\Component\Messenger\Stamp\TransportMessageIdStamp;
 use Vrok\MessengerReply\ReplyMiddleware;
-use Vrok\MessengerReply\ReplyStamp;
 use Vrok\MessengerReply\ReplyToStamp;
 
-class ReplyMiddlewareTest extends TestCase
+final class ReplyMiddlewareTest extends TestCase
 {
-    public function testConstruction()
+    public function testConstruction(): void
     {
         $bus = $this->createStub(MessageBusInterface::class);
         $logger = $this->createStub(LoggerInterface::class);
 
         $mw = new ReplyMiddleware($bus);
         $mw->setLogger($logger);
-
-        $this->assertInstanceOf(ReplyMiddleware::class, $mw);
     }
 
-    public function testIgnoresMessageWithoutReplyTo()
+    public function testIgnoresMessageWithoutReplyTo(): void
     {
-        $bus = $this->createStub(MessageBusInterface::class);
-        $logger = $this->createStub(LoggerInterface::class);
+        $bus = $this->createMock(MessageBusInterface::class);
+        $logger = $this->createMock(LoggerInterface::class);
 
         // it will not dispatch a new message but log a debug message showing
         // the middleware was called
@@ -55,19 +54,19 @@ class ReplyMiddlewareTest extends TestCase
         $res = $mw->handle($envelope, $stack);
 
         // it called the next middleware
-        $this->assertTrue($testMw->called);
+        self::assertTrue($testMw->called);
 
         // it returned the result of the next middleware
-        $this->assertNotNull($res->last(TransportMessageIdStamp::class));
+        self::assertNotNull($res->last(TransportMessageIdStamp::class));
 
         // it did _not_ attach a reply stamp
-        $this->assertNull($res->last(ReplyStamp::class));
+        self::assertNull($res->last(AmqpStamp::class));
     }
 
-    public function testRequiresHandledStamp()
+    public function testRequiresHandledStamp(): void
     {
-        $bus = $this->createStub(MessageBusInterface::class);
-        $logger = $this->createStub(LoggerInterface::class);
+        $bus = $this->createMock(MessageBusInterface::class);
+        $logger = $this->createMock(LoggerInterface::class);
 
         // it will not dispatch a new message but log a warning
         $bus->expects($this->never())
@@ -89,25 +88,24 @@ class ReplyMiddlewareTest extends TestCase
             ->with(new ReplyToStamp('output')), $stack);
 
         // it called the next middleware
-        $this->assertTrue($testMw->called);
+        self::assertTrue($testMw->called);
 
         // it returned the result of the next middleware
-        $this->assertNotNull($res->last(TransportMessageIdStamp::class));
+        self::assertNotNull($res->last(TransportMessageIdStamp::class));
 
         // it did _not_ attach a reply stamp
-        $this->assertNull($res->last(ReplyStamp::class));
+        self::assertNull($res->last(AmqpStamp::class));
     }
 
-    public function testRequiresResultObject()
+    public function testRequiresResultObject(): void
     {
-        $bus = $this->createStub(MessageBusInterface::class);
-        $logger = $this->createStub(LoggerInterface::class);
+        $bus = $this->createMock(MessageBusInterface::class);
+        $logger = $this->createMock(LoggerInterface::class);
 
         // it will not dispatch a new message but throw an error
         $bus->expects($this->never())
             ->method('dispatch');
-        $this->expectException(UnrecoverableMessageHandlingException::class,
-            'Result returned by handler TestHandler must be a serializable Message object to be a valid reply!');
+        $this->expectException(UnrecoverableMessageHandlingException::class);
 
         $mw = new ReplyMiddleware($bus);
         $mw->setLogger($logger);
@@ -119,14 +117,14 @@ class ReplyMiddlewareTest extends TestCase
             ->with(new HandledStamp(123, 'TestHandler')), $stack);
     }
 
-    public function testDispatchesReply()
+    public function testDispatchesReply(): void
     {
         // we cannot use a stub here because PHPUnit tries to mock the return type of dispatch(),
         // even when we use ->will($this->returnArgument(0)), and results in:
         // Class "Symfony\Component\Messenger\Envelope" is declared "final" and cannot be mocked.
         $bus = new TestBus();
 
-        $logger = $this->createStub(LoggerInterface::class);
+        $logger = $this->createMock(LoggerInterface::class);
 
         $mw = new ReplyMiddleware($bus);
         $mw->setLogger($logger);
@@ -144,24 +142,24 @@ class ReplyMiddlewareTest extends TestCase
             ->with(new HandledStamp(new TestReply(), 'TestHandler')), $stack);
 
         // the dispatch() method was called
-        $this->assertNotNull($bus->envelope);
+        self::assertNotNull($bus->envelope);
 
         // the routing key was added on the given reply
         $amqpStamp = $bus->envelope->last(AmqpStamp::class);
-        $this->assertNotNull($amqpStamp);
-        $this->assertSame('output', $amqpStamp->getRoutingKey());
+        self::assertNotNull($amqpStamp);
+        self::assertSame('output', $amqpStamp->getRoutingKey());
 
         // it dispatched our given handler result
-        $this->assertInstanceOf(TestReply::class, $bus->envelope->getMessage());
+        self::assertInstanceOf(TestReply::class, $bus->envelope->getMessage());
 
         // it called the next middleware
-        $this->assertTrue($testMw->called);
+        self::assertTrue($testMw->called);
 
         // it returned the result of the next middleware
-        $this->assertNotNull($res->last(TransportMessageIdStamp::class));
+        self::assertNotNull($res->last(TransportMessageIdStamp::class));
     }
 
-    public function testTransfersTaskIdentifier()
+    public function testTransfersTaskIdentifier(): void
     {
         // we cannot use a stub here because PHPUnit tries to mock the return type of dispatch(),
         // even when we use ->will($this->returnArgument(0)), and results in:
@@ -182,17 +180,17 @@ class ReplyMiddlewareTest extends TestCase
         $envelope = new Envelope($input);
         $replyTo = new ReplyToStamp('output');
 
-        $res = $mw->handle(
+        $mw->handle(
             $envelope->with($replyTo)
                 ->with(new HandledStamp(new TestReply(), 'TestHandler')),
             $stack
         );
 
         // it returned the given reply
-        $this->assertInstanceOf(TestReply::class, $bus->envelope->getMessage());
+        self::assertInstanceOf(TestReply::class, $bus->envelope->getMessage());
 
         // it copied to identifier properties from the input message
-        $this->assertSame('123', $bus->envelope->getMessage()->getIdentifier());
-        $this->assertSame('research', $bus->envelope->getMessage()->getTask());
+        self::assertSame('123', $bus->envelope->getMessage()->getIdentifier());
+        self::assertSame('research', $bus->envelope->getMessage()->getTask());
     }
 }
